refactor: #66 move aggregate factories into the Domain layer
The factory now lives in {Aggregates}/Domain/Factories beside the aggregate it
builds, so newFactory() no longer points from Domain into Infrastructure.

That was the only exemption on `domain does not depend on infrastructure`, and
it is gone: the rule now applies to every class in the layer with no list of
classes to keep in step. Nothing new reaches Domain by this, since the
aggregate root is already an Eloquent model.

Also correct two README claims: the registration bullet named two commands for
work a third does, and the paragraph on seeders and factories still described
the autoload map as it stood before database/seeders was mapped back.
